+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Allow `sentry-opentelemetry-agentless` and `sentry-opentelemetry-agentless-spring` to determine the SDK version to install ([#841](https://github.com/getsentry/sentry-android-gradle-plugin/pull/841))
+
 ### Fixes
 
 - Ensure `InjectSentryDebugMetaPropertiesTask` deletes the output folder before writing into it ([#838](https://github.com/getsentry/sentry-android-gradle-plugin/pull/838))